added case insensitivity to the uploaduserkey
Added case insensitivity to the uploaduserkey function to allow gpgkey and valid email to ignore case. This will allow it to import keys for email addresses that have varying capitalization.  This should work fine as email addresses are not case sensitive, and other email clients tend to handle this just fine.